Update README.md
# Eliza_Chat-Bot
This is my implementation of Eliza Chat Bot. I got the therapist patterns (intents.json) on Kaggle and adapted the code to run locally using python and Flask library.

To use it, you will need to install python 3, and the Flask library.

1 - Open Windows prompt command
2 - Go to the paste where you paste the file using as example CD C:\Users\eliza_aprimorada
3 - Digite python app.py
4 - The following message will appear:  
* Serving Flask app 'app'
 * Debug mode: off
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on http://127.0.0.1:5000
5 - Hold CTRL and right click on the server http address
6 - Eliza Chat Bot will open your web browser and run locally.
7 - If you wanna introduce more patterns just change the intents.json file

Have fun.